pyproject.toml
[tool.poetry]
name = "prediction"
version = "0.1.0"
description = "KGSA Prediction Service"
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "3.8.6"
numpy = "1.19.2"
pandas = "1.1.3"
pmdarima = "1.6.0"
fbprophet = "0.7.1"
holiday = "1.0.0"
jsonlines = "1.2.0"
PyYAML = "5.3.1"
sagemaker = ">=2.15.2"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp38-cp38-linux_x86_64.whl"}
smart-open = "^3.0.0"
kubernetes = "^12.0.0"
psycopg2 = "^2.8.6"
tabulate = "^0.8.7"
statsmodels = "0.12.2"
pony = "^0.7.14"
yoyo-migrations = "^7.2.1"
fsspec = "^0.8.5"
s3fs = "^0.5.2"
pydantic = "^1.8.2"
fastapi = {extras = ["all"], version = "^0.63.0"}
uvicorn = {extras = ["standard"], version = "^0.13.3"}
tensorflow = "^2.4.1"
ConfigArgParse = "^1.3"
backoff = "^1.10.0"
cloudpickle = "^1.6.0"
tensorboardX = "^2.2"
ray = "^1.13.0"
python-box = "^5.4.0"
lxml = "^4.6.3"
pystan = "2.19.1.1"
python-json-logger = "^2.0.2"
loguru = "^0.5.3"
aioredis = "<2"
anyscale = "0.5.1"
ddtrace = "^0.57.1"
simplejson = "^3.17.6"
MonthDelta = "^0.9.1"
scipy = "1.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
pre-commit = "^2.7.1"
black = "^20.8b1"
isort = "^5.6.4"
pytest-cov = "^2.10.1"
bandit = "^1.6.2"
jupyterlab = "^3.0.6"
moto = "^1.3.16"
flake8 = "^3.8.4"
mypy = "^0.800"
pytest-env = "^0.6.2"
httpie = "^2.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--strict-markers -m 'not slow'"
markers = [
    "ray",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
env = [
    "DATABASE_URL=postgres://postgres:local_db_password@localhost/postgres",
    "INPUT_PATH=./tests/resources/data/",
    "RUN_LOCAL=1",
    "LOCAL_RAY_CLUSTER=1",
    "JSON_LOGS=0"
]
